5 version

+++ src/main/java/com/dmitry/spreater/repos/MessageRepo.java
@@ -2,6 +2,7 @@ package com.dmitry.spreater.repos;
 
 import com.dmitry.spreater.domain.Message;
 import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
@@ -9,7 +10,7 @@ import java.util.List;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface MessageRepo extends CrudRepository<Message, Integer> {
+public interface MessageRepo extends CrudRepository<Message, Long> {
 
     List<Message> findByTag(String tag);
